+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * improved typing for reraise_as
 * added weak_refs to SingletonWithRegardsTo
 * added jump_to_directory
+* fixed get_size() to RuntimeError on PyPy instead of hanging